+++ codec/cbor/doc.go
@@ -1,2 +1,0 @@
-// Package cbor contains a codec implementation that marshals in CBOR format.
-package cbor