+++ CHANGELOG.md
@@ -10,7 +10,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 [keep a changelog]: https://keepachangelog.com/en/1.0.0/
 [semantic versioning]: https://semver.org/spec/v2.0.0.html
 
-## [Unreleased]
+## [0.8.0] - 2024-09-27
 
 ### Removed
 
@@ -186,6 +186,7 @@ different marshalers/unmarshalers provided by `google.golang.org/protobuf`.
 [0.7.3]: https://github.com/dogmatiq/marshalkit/releases/tag/v0.7.3
 [0.7.5]: https://github.com/dogmatiq/marshalkit/releases/tag/v0.7.5
 [0.7.6]: https://github.com/dogmatiq/marshalkit/releases/tag/v0.7.6
+[0.8.0]: https://github.com/dogmatiq/marshalkit/releases/tag/v0.8.0
 
 <!-- version template
 ## [0.0.1] - YYYY-MM-DD